refactor: 코드 분리 및 폴더명 변경
