+++ app/components/index.ts
@@ -1,13 +1,13 @@
-export { default as FloatingUpButton } from './common/FloatingUpButton';
-export { default as Header } from './common/Header';
-export { default as FilterChip } from './common/FilterChip/FilterChip';
-export { default as FilterChipsBox } from './common/FilterChip/FilterChipsBox';
-export { default as GlobalNavigationBar } from './common/GlobalNavigationBar';
-export { default as SNBFilterList } from './common/SideNavigationBar/SNBFilterList';
-export { default as SideNavigationBar } from './common/SideNavigationBar/SideNavigationBar';
+export { default as FloatingUpButton } from './layout/FloatingUpButton';
+export { default as Header } from './layout/Header';
+export { default as FilterChip } from './layout/FilterChip/FilterChip';
+export { default as FilterChipsBox } from './layout/FilterChip/FilterChipsBox';
+export { default as GlobalNavigationBar } from './layout/GlobalNavigationBar';
+export { default as SNBFilterList } from './layout/SideNavigationBar/SNBFilterList';
+export { default as SideNavigationBar } from './layout/SideNavigationBar/SideNavigationBar';
 
-export { default as ProductColor } from './Product/ProductColor';
-export { default as ProductInfo } from './Product/ProductInfo';
+export { default as ProductColor } from './product/ProductColor';
+export { default as ProductInfo } from './product/ProductInfo';
 
 export { default as DisplayProducts } from './DisplayProducts';
