+++ app/components/layout/GlobalNavigationBar.tsx
