+++ app/modules/hooks/productInfo.ts
@@ -1,0 +1,9 @@
+/**
+ * 상품 상세 조회 API
+ */
+export const getProductInfo = async (productId: string) => {
+  return fetch(
+    `http://makeup-api.herokuapp.com/api/v1/products/${productId}.json`,
+    { method: 'GET' },
+  ).then((res) => res.json());
+};
