+++ next.config.mjs
@@ -1,8 +1,6 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: false,
-  output: 'export', // 정적 웹 사이트 구축을 위한 설정
-  // trailingSlash: true, // 정적 웹 사이트 구축을 위한 설정
   images: {
     unoptimized: true, // 정적 웹 사이트 구축을 위한 설정
     remotePatterns: [
